requirements/analysis_requirements.txt
# Core framework dependencies
langgraph>=0.2.0
pydantic>=2.0.0

# LLM integration
google-genai>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # For Parquet support

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# File handling
openpyxl>=3.1.0  # For Excel support
sqlalchemy>=2.0.0  # For SQL support

# Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0 